backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-planning-backend"
version = "0.1.0"
description = "Local-first investment planner and portfolio analytics backend"
authors = [{name = "Financial Planning Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "sqlalchemy>=2.0.42",
    "alembic>=1.16.4",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "pandas>=2.3.1",
    "numpy>=2.3.2",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.1",
    "python-json-logger>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "httpx>=0.28.1",
    "ruff>=0.12.8",
    "mypy>=1.7.1",
    "black>=25.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
select = ["E", "F", "I", "N", "W", "UP"]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
